refactor: organize function components
