+++ src/components/Footer/style.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const FooterContainer = styled.footer`
+  width: 100%;
+  min-height: 48px;
+  background: #ffdc00;
+  color: #000;
+  display: flex;
+`;
+export const Icon = styled.div``;
+export const Notice = styled.img`
+  margin: 10px 6px 0 5px;
+  height: 22px;
+`;
+export const Announcement = styled.div`
+  font-size: 12px;
+  margin-top: 6px;
+`;
+export const Source = styled.p``;
+export const Warning = styled.p``;
