+++ src/components/GridModeContainer/GridModeContainer.js
@@ -1,0 +1,29 @@
+import React from 'react';
+import {
+  Container,
+  Item_1,
+  Item_2,
+  Item_3,
+  Item_4,
+  Item_5,
+  Item_6,
+  Item_7,
+  Item_8,
+} from './style';
+
+export default function GridModeContainer() {
+  return (
+    <>
+      <Container>
+        <Item_1></Item_1>
+        <Item_2></Item_2>
+        <Item_3></Item_3>
+        <Item_4></Item_4>
+        <Item_5></Item_5>
+        <Item_6></Item_6>
+        <Item_7></Item_7>
+        <Item_8></Item_8>
+      </Container>
+    </>
+  );
+}
