+++ src/components/GridModeContainer/style.js
@@ -1,0 +1,60 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  background: #2983de;
+  display: grid;
+  justify-content: center;
+  grid-template-columns: 20vw 20vw 20vw;
+  grid-template-rows: 20vw 20vw 20vw;
+  column-gap: 50px;
+  row-gap: 50px;
+  margin-top: 50px;
+`;
+export const Item_1 = styled.div`
+  background: url('/pokemon_photo/1.svg') center/cover no-repeat;
+  background-color: #2e4595;
+  width: 100%;
+  height: 100%;
+`;
+export const Item_2 = styled.div`
+  background: url('/pokemon_photo/2.svg') center/cover no-repeat;
+  background-color: #f8c145;
+  width: 100%;
+  height: 100%;
+`;
+export const Item_3 = styled.div`
+  background: url('/pokemon_photo/3.svg') center/cover no-repeat;
+  background-color: #2e4595;
+  width: 100%;
+  height: 100%;
+`;
+export const Item_4 = styled.div`
+  background: url('/pokemon_photo/4.svg') center/cover no-repeat;
+  background-color: #f8c145;
+  width: 100%;
+  height: 100%;
+`;
+export const Item_5 = styled.div`
+  background: url('/pokemon_photo/5.svg') center/cover no-repeat;
+  background-color: #2e4595;
+  width: 100%;
+  height: 100%;
+`;
+export const Item_6 = styled.div`
+  background: url('/pokemon_photo/6.svg') center/cover no-repeat;
+  background-color: #f8c145;
+  width: 100%;
+  height: 100%;
+`;
+export const Item_7 = styled.div`
+  background: url('/pokemon_photo/7.svg') center/cover no-repeat;
+  background-color: #2e4595;
+  width: 100%;
+  height: 100%;
+`;
+export const Item_8 = styled.div`
+  background: url('/pokemon_photo/8.svg') center/cover no-repeat;
+  background-color: #f8c145;
+  width: 100%;
+  height: 100%;
+`;
