+++ src/components/Header/Header.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Title, PokemonLogo } from './style';
+
+export default function Header() {
+  return (
+    <>
+      <Title>
+        <PokemonLogo />
+      </Title>
+    </>
+  );
+}
